tests/wpt/tests/credential-management/fedcm-multi-idp/single-get-during-onload.https.html

<!DOCTYPE html>
<title>Federated Credential Management API multi IDP single get during onload test.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>

<script type="module">
import {request_options_with_mediation_required,
        set_fedcm_cookie,
        fedcm_get_and_select_first_account} from '../support/fedcm-helper.sub.js';

promise_test(async t => {
  const window_loaded = new Promise(resolve => {
    window.addEventListener('load', async () => {
      await set_fedcm_cookie();
      const first_cred = fedcm_get_and_select_first_account(t,
          request_options_with_mediation_required());
      const cred = await first_cred;
      assert_equals(cred.token, "token");
      resolve();
    });
  });
  await window_loaded;
}, "Single `get` call during onload is allowed.");

</script>
